Check the validity of the service name
if user want to switch to `ubuntu-console` and type the wrong
service name `ubuntu-consile`, the command `sudo ros service enable
ubuntu-consile` run with no error. after the reboot, the console is
still the busybox console.

It is better to Warn user that the `ubuntu-consile` is no a valid
service name.

This patch also check the validity of the service name when disable and
delete service.

Signed-off-by: Wang Long <long.wanglong@huawei.com>